requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
djangorestframework_simplejwt
Markdown==3.3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
